updated db schema

+++ db_schema.sql
@@ -1,1 +1,1 @@
-CREATE TABLE arnu(receive INT NOT NULL, servicecode SMALLINT NOT NULL, transportmodecode VARCHAR(12), servicetype VARCHAR(40), contents TEXT);
+CREATE TABLE arnu(receive INT NOT NULL, servicecode INT NOT NULL, transportmodecode VARCHAR(12), servicetype VARCHAR(40), contents TEXT);
